requirements.txt
Ansible==2.7.2
Flask==1.0.2
PyYAML==3.13
firebase-admin==2.15.0
hashids==1.2.0